Add OsuMonitor class to monitor "osu!" process and avoid ungraceful termination of the program
Introduced a new class `OsuMonitor` in the `Osussist.src.osu.helpers` namespace to monitor the presence of the "osu!" process. The `OsuMonitor` class uses a `CancellationTokenSource` to periodically check if the "osu!" process is running, logging a message and exiting the application if the process is not found.

Updated `Main.cs` to instantiate and start the `OsuMonitor` using its `StartMonitoring` method. The `OsuMonitor` is stopped when the application exits by subscribing to the `AppDomain.CurrentDomain.ProcessExit` event. Added necessary `using` directives and ensured proper initialization and cleanup of the `OsuMonitor` in the `Main` method.